SauceDemo.Core project created. Solution is functional, tests are running. README updated.
