+++ SauceDemo.Tests/Config/appsettings___.json
@@ -1,0 +1,7 @@
+﻿{
+  "TestSettings": {
+    "Browser": "chrome",
+    "BaseUrl": "https://www.saucedemo.com/",
+    "WebDriverWaitSeconds": 10
+  }
+}
