+++ SauceDemo.Tests/SauceDemo.Tests.csproj
@@ -10,18 +10,22 @@
 	<ItemGroup>
 		<PackageReference Include="DotNetSeleniumExtras.WaitHelpers" Version="3.11.0" />
 		<PackageReference Include="FluentAssertions" Version="8.4.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.6" />
+		<!--<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.6" />-->
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.14.1" />
 		<PackageReference Include="NUnit" Version="4.3.2" />
 		<PackageReference Include="NUnit3TestAdapter" Version="5.0.0" />
-		<PackageReference Include="Selenium.Support" Version="4.33.0" />
+		<!--<PackageReference Include="Selenium.Support" Version="4.33.0" />
 		<PackageReference Include="Selenium.WebDriver" Version="4.33.0" />
 		<PackageReference Include="Selenium.WebDriver.ChromeDriver" Version="138.0.7204.4900" />
-		<PackageReference Include="Selenium.WebDriver.MSEdgeDriver" Version="136.0.3240.76" />
+		<PackageReference Include="Selenium.WebDriver.MSEdgeDriver" Version="136.0.3240.76" />-->
 		<PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
-		  <PrivateAssets>all</PrivateAssets>
-		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
+	</ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\SauceDemo.Core\SauceDemo.Core.csproj" />
+		<ProjectReference Include="..\SauceDemo.UI\SauceDemo.UI.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
@@ -29,7 +33,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<None Update="Config\appsettings.json">
+		<None Update="Config\appsettings___.json">
 			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
 		</None>
 	</ItemGroup>
