+++ SauceDemo.UI/Base/BasePage.cs
@@ -1,0 +1,16 @@
+﻿// <copyright file="BasePage.cs" company="PlaceholderCompany">
+// Copyright (c) PlaceholderCompany. All rights reserved.
+// </copyright>
+
+namespace SauceDemo.UI.Base
+{
+    /// <summary>
+    /// Abstract base class for all page objects.
+    /// Provides shared WebDriver instance and common UI interaction helpers.
+    /// </summary>
+    public abstract class BasePage
+    {
+            //        this.Driver = WebDriverFactory.Driver;
+            //this.Wait = new WebDriverWait(this.Driver, TimeSpan.FromSeconds(TestConfig.WebDriverWaitSeconds));
+    }
+}
