+++ SauceDemo.sln
@@ -10,6 +10,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		stylecop.json = stylecop.json
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SauceDemo.Core", "SauceDemo.Core\SauceDemo.Core.csproj", "{737A370C-BA09-42C3-8BFB-BA822986882F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SauceDemo.UI", "SauceDemo.UI\SauceDemo.UI.csproj", "{88BBEECE-56F4-4BB6-967D-A55FACCB3069}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -32,6 +36,30 @@ Global
 		{7B2D0B11-E17B-4231-B66F-1CA16F0C3E7F}.Release|x64.Build.0 = Release|Any CPU
 		{7B2D0B11-E17B-4231-B66F-1CA16F0C3E7F}.Release|x86.ActiveCfg = Release|Any CPU
 		{7B2D0B11-E17B-4231-B66F-1CA16F0C3E7F}.Release|x86.Build.0 = Release|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Debug|x64.Build.0 = Debug|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Debug|x86.Build.0 = Debug|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Release|x64.ActiveCfg = Release|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Release|x64.Build.0 = Release|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Release|x86.ActiveCfg = Release|Any CPU
+		{737A370C-BA09-42C3-8BFB-BA822986882F}.Release|x86.Build.0 = Release|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Debug|x64.Build.0 = Debug|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Debug|x86.Build.0 = Debug|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Release|Any CPU.Build.0 = Release|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Release|x64.ActiveCfg = Release|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Release|x64.Build.0 = Release|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Release|x86.ActiveCfg = Release|Any CPU
+		{88BBEECE-56F4-4BB6-967D-A55FACCB3069}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
